Move topic_coherece to gensim/models/_coherence + fix doc structure
