+++ docs/src/models/_coherence/aggregation.rst
@@ -1,0 +1,9 @@
+:mod:`models._coherence.aggregation` -- Aggregation module
+==========================================================
+
+.. automodule:: gensim.models._coherence.aggregation
+    :synopsis: Aggregation module
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
