+++ docs/src/models/_coherence/direct_confirmation_measure.rst
@@ -1,0 +1,9 @@
+:mod:`models._coherence.direct_confirmation_measure` -- Direct confirmation measure module
+==========================================================================================
+
+.. automodule:: gensim.models._coherence.direct_confirmation_measure
+    :synopsis: Direct confirmation measure module
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
