+++ docs/src/models/_coherence/indirect_confirmation_measure.rst
@@ -1,0 +1,9 @@
+:mod:`models._coherence.indirect_confirmation_measure` -- Indirect confirmation measure module
+==============================================================================================
+
+.. automodule:: gensim.models._coherence.indirect_confirmation_measure
+    :synopsis: Indirect confirmation measure module
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
