+++ docs/src/models/_coherence/probability_estimation.rst
@@ -1,0 +1,9 @@
+:mod:`models._coherence.probability_estimation` -- Probability estimation module
+================================================================================
+
+.. automodule:: gensim.models._coherence.probability_estimation
+    :synopsis: Probability estimation module
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
