+++ docs/src/models/_coherence/segmentation.rst
@@ -1,0 +1,9 @@
+:mod:`models._coherence.segmentation` -- Segmentation module
+============================================================
+
+.. automodule:: gensim.models._coherence.segmentation
+    :synopsis: Segmentation module
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
