+++ docs/src/models/_coherence/text_analysis.rst
@@ -1,0 +1,9 @@
+:mod:`models._coherence.text_analysis` -- Analyzing the texts of a corpus to accumulate statistical information about word occurrences
+======================================================================================================================================
+
+.. automodule:: gensim.models._coherence.text_analysis
+    :synopsis: Analyzing the texts of a corpus to accumulate statistical information about word occurrences
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
