--- docs/src/topic_coherence/aggregation.rst
@@ -1,9 +1,0 @@
-:mod:`topic_coherence.aggregation` -- Aggregation module
-========================================================
-
-.. automodule:: gensim.topic_coherence.aggregation
-    :synopsis: Aggregation module
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
