--- docs/src/topic_coherence/indirect_confirmation_measure.rst
@@ -1,9 +1,0 @@
-:mod:`topic_coherence.indirect_confirmation_measure` -- Indirect confirmation measure module
-============================================================================================
-
-.. automodule:: gensim.topic_coherence.indirect_confirmation_measure
-    :synopsis: Indirect confirmation measure module
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
