--- docs/src/topic_coherence/segmentation.rst
@@ -1,9 +1,0 @@
-:mod:`topic_coherence.segmentation` -- Segmentation module
-==========================================================
-
-.. automodule:: gensim.topic_coherence.segmentation
-    :synopsis: Segmentation module
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
