+++ gensim/models/_coherence/segmentation.py
